// contracts/generated/typescript/models/AssetRevision.ts
export interface AssetRevision {
    /**
     *
     * @type {string}
     * @memberof AssetRevision
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof AssetRevision
     */
    assetId: string;
    /**
     *
     * @type {string}
     * @memberof AssetRevision
     */
    contentHash: string;
    /**
     *
     * @type {string}
     * @memberof AssetRevision
     */
    mimeType: string;
    /**
     *
     * @type {number}
     * @memberof AssetRevision
     */
    sizeBytes: number;
    /**
     *
     * @type {number}
     * @memberof AssetRevision
     */
    width: number;
    /**
     *
     * @type {number}
     * @memberof AssetRevision
     */
    height: number;
    /**
     *
     * @type {Date}
     * @memberof AssetRevision
     */
    createdAt: Date;
}